fix windows tests
