+++ pytest.ini
@@ -1,4 +1,4 @@
 [pytest]
-; python_paths = src galaxy-integrations-python-api/src ; dev env, to run vs build package see `tasks.py`
+python_paths = src
 testpaths = tests
 addopts = --flakes
