+++ src/consts.py
@@ -59,6 +59,7 @@ else:
     raise PlatformNotSupported('GOG Galaxy 2.0 supports only Windows and macos for now')
 
 IS_WINDOWS = HP.WINDOWS == CURRENT_SYSTEM
+IS_MAC = HP.MAC == CURRENT_SYSTEM
 
 if platform.machine().endswith('64'):
     CURRENT_BITNESS = BITNESS.B64
